Feature: Implementación sugerencias en asesoria con tutor Se ajusta double  a big decimal Implementación response codes Ajuste en caso de uso para validacion y codigos para BusinessException Estandarizacion de respuesta con ApiResponse Implementación de TechnicalException Creación de constantes para string quemados

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/bancolombia/r2dbc/entity/LoanTypeEntity.java
@@ -5,6 +5,8 @@ import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder(toBuilder = true)
@@ -24,10 +26,10 @@ public class LoanTypeEntity {
     private Double minAmount;
 
     @Column("monto_maximo")
-    private Double maxAmount;
+    private BigDecimal maxAmount;
 
     @Column("tasa_interes")
-    private Double interestRate;
+    private BigDecimal interestRate;
 
     @Column("validacion_automatica")
     private Boolean automaticValidation;
